Bumps the Rust toolchain version (#687)
This has a few useful effects:

- Hopefully fixes a recent OSS-Fuzz failure to compile.
- Decreases binary size by 0.8 kB.
- Fixes a warning by `cargo audit`.

We also bump our minimum supported Rust version to what it actually is.
We had an API call for `BTreeSet` that needs Rust 1.66. If you use an
older version, you have to patch OpenSK and replace `pop_first` with an
iterator that takes the first element, like we did before `pop_first`
was available.